% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLP1R.R
\docType{data}
\name{GLP1R}
\alias{GLP1R}
\title{GLP1R summary data included in the package}
\description{
GLP1R summary data included in the package
}
\references{
\url{https://gtexportal.org/home/tissueSummaryPage}
}
\keyword{data}
