% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop-coloc.R
\name{prop.coloc}
\alias{prop.coloc}
\title{A frequentist test of proportional colocalization after selecting relevant genetic variants}
\usage{
prop.coloc(
  b1,
  se1,
  b2,
  se2,
  n,
  ld,
  tau = NULL,
  alpha = NULL,
  prune = NULL,
  J = NULL,
  figs = NULL,
  traits = NULL,
  seed = 100
)
}
\arguments{
\item{b1}{beta coefficients for trait 1}

\item{se1}{standard errors for trait 1}

\item{b2}{beta coefficients for trait 2}

\item{se2}{standard errors for trait 2}

\item{n}{sample size}

\item{ld}{genetic variant correlation matrix}

\item{tau}{(optional) correlation between trait 1 and trait 2; if traits are measured in separate samples, then this should be set to 0 (default value is 0)}

\item{alpha}{(optional) nominal size of the conditional proportional colocalization test; if unspecified, p-value is reported to 2 decimal places}

\item{prune}{(optional) R^2 pruning threshold for variant correlations; if unspecified (default value is R^2 = 0.6)}

\item{J}{(optional) the top J variants for at least one trait are used to fit multivariable trait-variant linear models (default value is J=10)}

\item{figs}{(optional) logical: if \code{TRUE}, return plots of the univariable and fitted multivariable variant--trait associations of top J variants for each trait (default value is \code{FALSE})}

\item{traits}{(optional) a character vector of the two trait names}

\item{seed}{(optional) the seed for random sampling involved in computing conditional critical values for the prop-coloc-cond p-value (default value is 100)}
}
\value{
Output is a list containing:
\itemize{
\item \code{p_cond} \cr
p-value of prop-coloc-cond to two decimal places, or if \code{alpha} specified then logical \code{TRUE}/\code{FALSE} for whether the proportional colocalization null hypothesis is rejected (if \code{TRUE}) or not rejected (if \code{FALSE}) at the specified \code{alpha} significance threshold. Null hypothesis is that the beta-coefficients are proportional (this represents proportional colocalization), rejection indicates failure to proportionately colocalize.
\item \code{eta_cond} \cr
proportionality constant based on lead variants for each trait
\item \code{LM_cond} \cr
p-value of Lagrange Multiplier test based on lead variants for each trait. Null hypothesis is that the proportionality constant is zero, rejection indicates proportionality constant is non-zero. Colocalization is only meaningful if the proportionality constant is non-zero.
\item \code{p_full} \cr
p-value of prop-coloc-full test based on top J variants for each trait. Null hypothesis is that the beta-coefficients are proportional (this represents proportional colocalization), rejection indicates failure to colocalize.
\item \code{eta_full} \cr
proportionality constant based on top J variants for each trait
\item \code{LM_full} \cr
p-value of Lagrange Multiplier test based on top J variants for each trait. Null hypothesis is that the proportionality constant is zero, rejection indicates proportionality constant is non-zero. Colocalization is only meaningful if the proportionality constant is non-zero.
\item \code{Q} \cr
naive test statistic for proportional colocalization hypothesis from prop-coloc-cond
\item \code{alpha} \cr
nominal size of the conditional proportional colocalization test (if specified)
\item \code{convergence} \cr
whether uniroot function converged to find a conditional critical value for the prop-coloc-cond test
\item \code{top} \cr
two most relevant variants selected for the prop-coloc-cond test
\item \code{variants} \cr
set of variants used for the prop-coloc-full test
\item \code{fig_uni} \cr
plot of the univariable variant--trait associations of top J variants for each trait (if \code{figs} is specified and \code{TRUE})
\item \code{fig_multi} \cr
plot of the fitted multivariable variant--trait associations of top J variants for each trait (if \code{figs} is specified and \code{TRUE})
}
}
\description{
A proportional colocalization test that accounts for uncertainty in variant selection using summary data.
}
\details{
The method computes two different proportional colocalization tests.  \cr
\cr
First, prop-coloc-cond is a conditional test based on lead variants and it accounts for uncertainty in variant selection. By contrast, prop-coloc-full tests the proportional colocalization hypothesis using a larger set of variant associations (by default, the top 10 strongest variant associations for each trait).  \cr
\cr
Compared with prop-coloc-full, the prop-coloc-cond test is shown to achieve good type I error control, and be less sensitive to measurement errors in estimated genetic correlations. However, prop-coloc-cond and prop-coloc-full test slightly different null hypotheses; whereas prop-coloc-cond focuses on the evidence of lead variants, prop-coloc-full considers the evidence across more variants.  \cr
\cr
We note that an optional input of "tau", which is the correlation between the two traits, is typically not provided in usual genetic association summary data. Therefore, sensitivity of analyses to the choice of tau is recommended.
}
\examples{
res <- prop.coloc(b1=GLP1R$thyroid$beta, se1=GLP1R$thyroid$se, b2=GLP1R$lung$beta, se2=GLP1R$lung$se, ld=GLP1R$ld, n=838, alpha=0.05)

}
\references{
A frequentist test of proportional colocalization after selecting relevant genetic variants. Preprint.
}
\author{
Stephen Burgess and Ashish Patel
}
